Crate telegram_bot

Reexports

pub use connector::*;

pub use prelude::*;

pub use types::*;

Modules

connector
IO backend.

prelude
Telegram bot prelude.

types
Telegram bot types.

Structs

Api
Main type for sending requests to the Telegram bot API.

Config
Configuration for an Api.

Error
The Error type.

TelegramFuture
Represent a future that resolves into Telegram API response.

UpdatesStream
This type represents stream of Telegram API updates and uses long polling method under the hood.

Enums

ErrorKind
The kind of an error.