Enum telegram_bot::Error

pub enum Error {
    Http(Error),
    Io(Error),
    JsonDecode(DecoderError),
    JsonEncode(EncoderError),
    Api(String),
    InvalidState(String),
    InvalidTokenFormat(ParseError),
    InvalidEnvironmentVar(VarError),
    InvalidPath(String),
}

Telegram-Bot Error: Anything that may fail (HTTP, JSON, ...)

Variants

Http

HTTP related error

Io

IO related error (mainly reading the http result)

JsonDecode

Error while decoding JSON data

JsonEncode

Error while encoding JSON data

Api

Telegram server reponsded with an error + description

InvalidState

This should never happen (it possibly could if the telegram servers would respond with garbage)

InvalidTokenFormat

Occurs, if the given bot token would not result in a valid request URL.

InvalidEnvironmentVar

The given environment variable could not be fetched.

InvalidPath

The given path is not valid.