Enum telegram_bot::ListeningAction

pub enum ListeningAction {
    Continue,
    Stop,
}

A listening handler returns this type to signal the listening-method either to stop or to continue. If a handler returns Stop, the update it was passed counts as "handled" and won't be handled again.

Variants

Continue

Stop