Enum telegram_bot::ListeningMethod

pub enum ListeningMethod {
    LongPoll(Option<Integer>),
}

Different method how to listen for new updates. Currently LongPoll is the only method supported by this library. The Telegram API offers a webhook method which is not yet implemented here.

Variants
LongPoll(Option<Integer>)