Module telegram_bot::prelude

Telegram bot prelude.

This module re-exports request builder traits from telegram-bot-raw.

Traits

CanDeleteMessage
Delete messages..

CanEditMessageCaption
Edit captions of messages sent by the bot.

CanEditMessageReplyMarkup
Edit reply markup of messages sent by the bot.

CanEditMessageText
Edit text of messages sent by the bot.

CanForwardMessage
Forward message.

CanGetChat
Get up to date information about the chat.

CanGetChatAdministrators
Get a list of administrators in a chat.

CanGetChatMemberForChat
Get information about a member of a chat.

CanGetChatMemberForUser
Get information about a member of a chat.

CanGetChatMembersCount
Get the number of members in a chat.

CanGetFile
Get basic info about a file and prepare it for downloading.

CanGetUserProfilePhotos
Get a list of profile pictures for a user.

CanKickChatMemberForChat
Kick a user from a group or a supergroup.

CanKickChatMemberForUser
Kick a user from a group or a supergroup.

CanLeaveChat
Leave a group, supergroup or channel.

CanReplySendContact
Reply with phone contact.

CanReplySendLocation
Reply with point on the map.

CanReplySendMessage
Reply with text message.

CanReplySendVenue
Reply with information about a venue.

CanSendChatAction
Send action to a chat.

CanSendContact
Send phone contact.

CanSendLocation
Send point on the map.

CanSendMessage
Send text message.

CanSendVenue
Send information about a venue.

CanUnbanChatMemberForChat
Unban a previously kicked user in a supergroup or channel.

CanUnbanChatMemberForUser
Unban a previously kicked user in a supergroup or channel.

ToReplyRequest
Use this trait to convert a complex type to corresponding request and reply to the message.

ToRequest
Use this trait to convert a complex type to corresponding request and send it to the chat.