Struct telegram_bot::Bot

pub struct Bot {
    // some fields omitted
}

The Bot. Main type for doing nearly everything.

Methods

impl Bot

fn new(token: String) -> Bot

Creates a new bot with the given token. If the token is completely invalid (resulting in an invalid API-URL), the function will panic. However, the function will not check if the given token is a valid Telegram token. You can call get_me to execute a test request.

fn get_me(&mut self) -> Result<User>

Corresponds to the "getMe" method of the API.

fn send_message(&mut self, chat_id: Integer, text: String, disable_web_page_preview: Option<bool>, reply_to_message_id: Option<Integer>, reply_markup: Option<ReplyMarkup>) -> Result<Message>

Corresponds to the "sendMessage" method of the API.

fn forward_message(&mut self, chat_id: Integer, from_chat_id: Integer, message_id: Integer) -> Result<Message>

Corresponds to the "forwardMessage" method of the API.

fn send_location(&mut self, chat_id: Integer, latitude: Float, longitude: Float, reply_to_message_id: Option<Integer>, reply_markup: Option<ReplyMarkup>) -> Result<Message>

Corresponds to the "sendLocation" method of the API.

fn send_chat_action(&mut self, chat_id: Integer, action: ChatAction) -> Result<bool>

Corresponds to the "sendChatAction" method of the API.

fn get_user_profile_photos(&mut self, user_id: Integer, offset: Option<Integer>, limit: Option<Integer>) -> Result<UserProfilePhotos>

Corresponds to the "getUserProfilePhotos" method of the API.

fn get_updates(&mut self, offset: Option<Integer>, limit: Option<Integer>, timeout: Option<Integer>) -> Result<Vec<Update>>

Corresponds to the "getUpdates" method of the API.

Note: The method will not set the offset parameter on its own. To receive updates in a more high level way, see long_poll.

fn set_webhook<U: IntoUrl>(&mut self, url: Option<U>) -> Result<bool>

fn long_poll<H>(&mut self, timeout: Option<Integer>, handler: H) -> Result<()> where H: FnMut(&mut Bot, Update) -> Result<()>

Receive and handle updates via "getUpdates".

This method will repeatedly call get_updates to receive new updates. It will then call the given handler for every update and increase the update offset accordingly, so the handler will never be called with the same update twice. The timeout parameter influences how long (in seconds) each poll may last. Defaults to 30. The handler gets a mutable reference to the bot since borrowing it from the outer scope won't work. When the handler returns an Err value the bot will stop listening for updates and long_poll will return the Error. If you want to stop listening you can just return Error::UserInterrupt.

Note: If the bot is restarted, but the last received updates are not yet confirmed (the last poll was not empty), there will be some duplicate updates.