Struct telegram_bot::UpdatesStream

#[must_use = "streams do nothing unless polled"]pub struct UpdatesStream { /* fields omitted */ }

This type represents stream of Telegram API updates and uses long polling method under the hood.

Implementations

impl UpdatesStream

pub fn new(api: &Api) -> Self

create a new UpdatesStream instance.

pub fn timeout(&mut self, timeout: Duration) -> &mut Self

Set timeout for long polling requests, this corresponds with timeout field in getUpdates method, also this stream sets an additional request timeout for timeout + 1 second in case of invalid Telegram API server behaviour.

Default timeout is 5 seconds.

pub fn allowed_updates(
&mut self,
allowed_updates: &[AllowedUpdate]
) -> &mut Self

Set allowed updates to receive, this corresponds with allowed_updates field in getUpdates method. List the types of updates you want your bot to receive. For example, specify [“message”, “edited_channel_post”, “callback_query”] to only receive updates of these types. See Update for a complete list of available update types. Specify an empty list to receive all updates regardless of type (default). If not specified, the previous setting will be used.

Please note that this parameter doesn't affect updates created before the call to the getUpdates, so unwanted updates may be received for a short period of time.

pub fn limit(&mut self, limit: Integer) -> &mut Self

Set limits the number of updates to be retrieved, this corresponds with limit field in getUpdates method. Values between 1—100 are accepted.

Defaults to 100.

pub fn error_delay(&mut self, delay: Duration) -> &mut Self

Set a delay between erroneous request and next request. This delay prevents busy looping in some cases.

Default delay is 500 ms.