Enum telegram_bot::types::Chat

pub enum Chat {
    Private {
        id: Integer,
        first_name: String,
        last_name: Option<String>,
        username: Option<String>,
    },
    Group {
        id: Integer,
        title: String,
        is_supergroup: bool,
    },
    Channel {
        id: Integer,
        title: String,
        name: Option<String>,
    },
}

Either a Private or a Group or a Channel. Used in "chat" field of Message. Has some useful methods for less typing.

Variants

Private
Fields

id

first_name

last_name

username

Group
Fields

id

title

is_supergroup

Channel
Fields

id

title

name

Methods

impl Chat

fn id(&self) -> Integer

Returns the chat id, which is needed to send messages.

fn is_user(&self) -> bool

Returns if the Chat is a User

fn is_group(&self) -> bool

Returns if the Chat is a Group

fn is_supergroup(&self) -> bool

Returns if the Chat is a SuperGroup

fn is_channel(&self) -> bool

Returns if the Chat is a Channel

fn to_user(&self) -> Option<User>