Struct telegram_bot::types::ChannelPost

pub struct ChannelPost {
    pub id: MessageId,
    pub date: i64,
    pub chat: Channel,
    pub forward: Option<Forward>,
    pub reply_to_message: Option<Box<MessageOrChannelPost, Global>>,
    pub edit_date: Option<i64>,
    pub kind: MessageKind,
}

This object represents a channel message.

Fields
id: MessageId

Unique message identifier inside this chat.

date: i64

Date the message was sent in Unix time.

chat: Channel

Conversation the message belongs to.

forward: Option<Forward>

Information about the original message.

reply_to_message: Option<Box<MessageOrChannelPost, Global>>

For replies, the original message. Note that the Message object in this field will not contain further reply_to_message fields even if it itself is a reply.

edit_date: Option<i64>

Date the message was last edited in Unix time.

kind: MessageKind

Kind of the message.