Struct telegram_bot::types::Video

pub struct Video {
    pub file_id: String,
    pub width: i64,
    pub height: i64,
    pub duration: i64,
    pub thumb: Option<PhotoSize>,
    pub mime_type: Option<String>,
    pub file_size: Option<i64>,
}

This object represents a video file.

Fields
file_id: String

Unique identifier for this file.

width: i64

Video width as defined by sender.

height: i64

Video height as defined by sender.

duration: i64

Duration of the video in seconds as defined by sender.

thumb: Option<PhotoSize>

Video thumbnail.

mime_type: Option<String>

Mime type of a file as defined by sender.

file_size: Option<i64>

File size.