Trait telegram_bot::util::messages::MessageGetFiles

pub trait MessageGetFiles {
    fn get_files<'a>(&'a self) -> Option<Vec<GetFile>>;
}

A trait to obtain GetFile requests from a message.

Many message kinds such as Sticker return a single GetFile. Message kinds like Photo might return more if an album is posted. A video, video note or document returns any thumbnail as well.

Required methods

fn get_files<'a>(&'a self) -> Option<Vec<GetFile>>

Obtain files from a message if available.

Loading content...