Crate telegram_bot

This crate helps writing bots for the messenger Telegram. See readme for details.

Re-exports

pub use prelude::*;

pub use types::*;

Modules

connector
Connector with hyper backend.

prelude
Telegram bot prelude.

types
Telegram bot types.

util
Traits that extends raw types functionality

Macros

reply_markup
Create a reply markup.

Structs

Api
Main type for sending requests to the Telegram bot API.

Error

UpdatesStream
This type represents stream of Telegram API updates and uses long polling method under the hood.