Module telegram_bot::prelude

Telegram bot prelude.

This module re-exports request builder traits from telegram-bot-raw.

Re-exports

pub use crate::util::messages::MessageGetFiles;

pub use crate::util::messages::MessageText;

Traits

CanAnswerCallbackQuery
Send answers to callback queries sent from inline keyboards.

CanAnswerInlineQuery

CanDeleteMessage
Delete messages..

CanEditMessageCaption
Edit captions of messages sent by the bot.

CanEditMessageLiveLocation
Edit live location messages sent by the bot.

CanEditMessageReplyMarkup
Edit reply markup of messages sent by the bot.

CanEditMessageText
Edit text of messages sent by the bot.

CanExportChatInviteLink
Export chat invite link.

CanForwardMessage
Forward message.

CanGetChat
Get up to date information about the chat.

CanGetChatAdministrators
Get a list of administrators in a chat.

CanGetChatMemberForChat
Get information about a member of a chat.

CanGetChatMemberForUser
Get information about a member of a chat.

CanGetChatMembersCount
Get the number of members in a chat.

CanGetFile
Get basic info about a file and prepare it for downloading.

CanGetUserProfilePhotos
Get a list of profile pictures for a user.

CanKickChatMemberForChat
Kick a user from a group or a supergroup.

CanKickChatMemberForUser
Kick a user from a group or a supergroup.

CanLeaveChat
Leave a group, supergroup or channel.

CanPinMessage

CanReplySendAudio
Can reply with an audio

CanReplySendContact
Reply with phone contact.

CanReplySendDocument
Can reply with a document

CanReplySendLocation
Reply with point on the map.

CanReplySendMessage
Reply with text message.

CanReplySendPhoto
Can reply with an photo

CanReplySendPoll

CanReplySendVenue
Reply with information about a venue.

CanReplySendVideo
Can reply with an video

CanSendAudio
Send an audio

CanSendChatAction
Send action to a chat.

CanSendContact
Send phone contact.

CanSendDocument
Send a document

CanSendLocation
Send point on the map.

CanSendMessage
Send text message.

CanSendPhoto
Send an photo

CanSendPoll
Send message with a poll.

CanSendVenue
Send information about a venue.

CanSendVideo
Send an video

CanStopMessageLiveLocation
Stop updating a live location message sent by the bot.

CanStopPoll
Stop a poll which was sent by the bot.

CanUnbanChatMemberForChat
Unban a previously kicked user in a supergroup or channel.

CanUnbanChatMemberForUser
Unban a previously kicked user in a supergroup or channel.

CanUnpinMessage

ToReplyRequest
Use this trait to convert a complex type to corresponding request and reply to the message.

ToRequest
Use this trait to convert a complex type to corresponding request and send it to the chat.