Module telegram_bot::types

Telegram bot types.

Reexports

pub use requests::*;

Modules

requests
Telegram Bot API methods.

Structs

Audio
This object represents an audio file to be treated as music by the Telegram clients.

Channel
This object represents a channel.

ChannelId
Unique channel identifier.

ChatId
Unique chat identifier.

Contact
This object represents a phone contact.

DetachedRequest
Partially serialized request.

Document
This object represents a general file (as opposed to photos, voice messages and audio files).

File
This object represents a file ready to be downloaded. The file can be downloaded via the link https://api.telegram.org/file/bot<token>/<file_path>. It is guaranteed that the link will be valid for at least 1 hour.

FileRef
Unique file identifier reference.

ForceReply

Forward
Information about the original message.

Group
This object represents a group.

GroupId
Unique group identifier.

IdResponse
Read T from the Telegram server and returns the same type.

InlineKeyboardMarkup

KeyboardButton

Location
This object represents a point on the map.

Message
This object represents a message.

MessageEntity
This object represents one special entity in a text message. For example, hashtags, usernames, URLs, etc.

MessageId
Unique message identifier inside a chat.

PhotoSize
This object represents one size of a photo or a file / sticker thumbnail.

ReplyKeyboardMarkup

ReplyKeyboardRemove

Sticker
This object represents a sticker.

Supergroup
This object represents a supergroup.

SupergroupId
Unique supergroup identifier.

TrueToUnitResponse
Read True from the Telegram server and returns ().

Update
This object represents an incoming update.

User
This object represents a Telegram user or bot.

UserId
Unique user identifier.

Venue
This object represents a venue.

Video
This object represents a video file.

Voice
This object represents a voice note.

Enums

Chat
This object represents a chat.

ChatAction
Strongly typed ChatAction. Instead of passing a String to the chat_action method, this is used.

ChatRef
Unique identifier for the target chat or username of the target channel (in the format @channelusername)

ForwardFrom
Information about the source of the original message.

MessageEntityKind
Kind of the entity.

MessageKind
Kind of the message.

ParseMode
Strongly typed ParseMode. See documentation for details.

ReplyMarkup

UpdateKind
Kind of the incoming update.

Traits

Request
Request corresponds to the specific Telegram API method.

Response
This trait represent how to parse response from the Telegram server.

ToChatRef
Get ChatRef from the type reference.

ToFileRef
Get FileRef from the type reference.

ToMessageId
Get MessageId from the type reference.

ToSourceChat
Get source ChatId from the type reference.

ToUserId
Get UserId from the type reference.

Type Definitions

Float
The Telegram Float, currently f32.

Integer
The Telegram Integer, currently i64.