Module telegram_bot::types::reply_markup

Structs

ForceReply
Upon receiving a message with this object, Telegram clients will display a reply interface to the user (act as if the user has selected the bot‘s message and tapped ’Reply'). This can be extremely useful if you want to create user-friendly step-by-step interfaces without having to sacrifice privacy mod

InlineKeyboardButton
This object represents one button of an inline keyboard.

InlineKeyboardMarkup
This object represents an inline keyboard that appears right next to the message it belongs to.

KeyboardButton
This object represents one button of the reply keyboard.

ReplyKeyboardMarkup
This object represents a custom keyboard with reply options.

ReplyKeyboardRemove
Upon receiving a message with this object, Telegram clients will remove the current custom keyboard and display the default letter-keyboard. By default, custom keyboards are displayed until a new keyboard is sent by a bot. An exception is made for one-time keyboards that are hidden immediately after the user presses a button (see ReplyKeyboardMarkup).

Enums

InlineKeyboardButtonKind

ReplyMarkup