Struct telegram_bot::types::Message

pub struct Message {
    pub id: MessageId,
    pub from: Option<User>,
    pub date: i64,
    pub chat: Chat,
    pub forward: Option<Forward>,
    pub reply_to_message: Option<Box<Message>>,
    pub edit_date: Option<i64>,
    pub kind: MessageKind,
}

This object represents a message.

Fields
id: MessageId

Unique message identifier inside this chat.

from: Option<User>

Sender, can be empty for messages sent to channels.

date: i64

Date the message was sent in Unix time.

chat: Chat

Conversation the message belongs to.

forward: Option<Forward>

Information about the original message.

reply_to_message: Option<Box<Message>>

For replies, the original message. Note that the Message object in this field will not contain further reply_to_message fields even if it itself is a reply.

edit_date: Option<i64>

Date the message was last edited in Unix time.

kind: MessageKind

Kind of the message.